refactor: remove dataloader factory and mock dataloader interface
